refactor: nodes module
